third_party/WebKit/LayoutTests/fast/frames/seamless/seamless-min-max-expected.txt

Test that seamless iframes respect min/max height/width CSS styles.
PASS successfullyParsed is true

TEST COMPLETE
PASS window.getComputedStyle(iframe1).width is "300px"
PASS window.getComputedStyle(iframe1).height is "300px"
PASS window.getComputedStyle(iframe2).width is "50px"
PASS window.getComputedStyle(iframe2).height is "50px"
